Final tidy up in tests.
