+++ tests/api/test_api.py
@@ -1,70 +1,8 @@
 """Tests for the API."""
 
 from datetime import date
-from unittest.mock import patch
 
 import pytest
-from ic_data_repo.permissions import deposit_action
-from invenio_access.permissions import ActionUsers
-from invenio_oauth2server.models import Token
-from invenio_oauth2server.proxies import current_oauth2server
-
-
-@pytest.fixture
-def user(UserFixture, app, db):
-    """An initialised user."""
-    u = UserFixture(
-        email="foo@bar.com",
-        password="password",
-    )
-    u.create(app, db)
-    return u
-
-
-@pytest.fixture
-def user_client(user, client):
-    """A client logged in as the user fixture."""
-    return user.login(client)
-
-
-@pytest.fixture
-def user_depositor(user, db):
-    """Give the user fixture permission to create deposits."""
-    db.session.add(ActionUsers.allow(deposit_action, user_id=user.id))
-    return user
-
-
-@pytest.fixture(autouse=True)
-def mock_manifest():
-    """Mock manifest to always return a value for theme.css."""
-    with patch("flask_webpackext.manifest.JinjaManifest.__getitem__") as mock:
-        mock.return_value = '<link rel="stylesheet" href="/static/dist/theme.css">'
-        yield mock
-
-
-@pytest.fixture
-def browser_headers(user_client):
-    """Headers for web browser requests, including CSRF token."""
-    csrf_token = user_client.get_cookie("csrftoken").value
-    if not csrf_token:
-        raise ValueError("CSRF token not found in cookies")
-    return {
-        "Content-Type": "application/json",
-        "X-CSRFToken": csrf_token,
-    }
-
-
-@pytest.fixture
-def api_headers(db, user_client, user_depositor):
-    """Headers for API requests, including API token."""
-    db.session.add(ActionUsers.allow(deposit_action, user_id=user_depositor.id))
-    scopes = [s[0] for s in current_oauth2server.scope_choices()]
-    token = Token.create_personal("test_token", user_depositor.id, scopes=scopes)
-    db.session.commit()
-    return {
-        "Authorization": f"Bearer {token.access_token}",
-        "Content-Type": "application/json",
-    }
 
 
 @pytest.fixture
